+++ libft/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: wding-ha <wding-ha@student.42kl.edu.my>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/12 11:07:04 by wding-ha          #+#    #+#             */
-/*   Updated: 2021/05/17 12:12:00 by wding-ha         ###   ########.fr       */
+/*   Updated: 2021/06/01 16:56:17 by wding-ha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
 	size_t	i;
